+++ app/Policies/ServerPolicy.php
@@ -24,11 +24,14 @@
 
 namespace Pterodactyl\Policies;
 
+use Cache;
+use Carbon;
 use Pterodactyl\Models\User;
 use Pterodactyl\Models\Server;
 
 class ServerPolicy
 {
+
     /**
      * Create a new policy instance.
      *
@@ -53,7 +56,13 @@ class ServerPolicy
             return true;
         }
 
-        return $user->permissions()->server($server)->permission($permission)->exists();
+        $permissions = Cache::remember('ServerPolicy.' . $user->uuid . $server->uuid, Carbon::now()->addSeconds(10), function () use ($user, $server) {
+            return $user->permissions()->server($server)->get()->transform(function ($item) {
+                return $item->permission;
+            })->values();
+        });
+
+        return ($permissions->search($permission, true) !== false);
     }
 
     /**