+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project follows [Semantic Versioning](http://semver.org) guidelines.
 ### Fixed
 * `[pre.3]` — Fixes bug in cache handler that doesn't cache against the user making the request. Would have allowed for users to access servers not belonging to themselves in production.
 
+### Added
+* New cache policy for ServerPolicy to avoid making 15+ queries per page load when confirming if a user has permission to perform an action.
+
 ## v0.6.0-pre.3 (Courageous Carniadactylus)
 ### Fixed
 * `[pre.2]` — Fixes bug where servers could not be manually deployed to nodes due to a broken SQL call.